refactor(core): extract control wire protocol to sigil-core::control_proto (#55)
Move the serializable control-socket wire types (Request/Response/payloads/
DoctorAiGuardReport + sub-structs) and the pure socket-path resolver from
sigil-agent into sigil-core::control_proto, so sigil-mcp can share the wire
contract without depending on the agent crate. cfg(operator-cli) gates dropped
in core (types always exist); agent keeps a not(operator-cli) fallback arm and
re-exports the types. No behavior change.
